fix bug where data was changing to np.float64 (from float32), this results in a 1.2x speedup. Added rerun_pyrlcade_h5py.py for rerunning runs wiht different seed

+++ pyrlcade/state/replay_buff.py
@@ -5,7 +5,7 @@ class replay_buff(object):
     def init(self,state_size,max_buf_size,debug_level=1):
         self.s = np.zeros((state_size,max_buf_size),dtype=np.float32)
         self.s2 = np.zeros((state_size,max_buf_size),dtype=np.float32)
-        self.a = np.zeros((max_buf_size))
+        self.a = np.zeros((max_buf_size),dtype=np.float32)
         self.r = np.zeros((max_buf_size),dtype=np.float32)
         self.term = np.zeros((max_buf_size),dtype=np.int)
         self.max_buf_size = max_buf_size
